// test/integration/test/ports/engines/queries/filters/field_reference/having_filter/_fixture/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly TestModel: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly string: CodecTypes['pg/text@1']['output'] | null;
      readonly string2: CodecTypes['pg/text@1']['output'] | null;
      readonly int: CodecTypes['pg/int4@1']['output'] | null;
      readonly int2: CodecTypes['pg/int4@1']['output'] | null;
      readonly bInt: CodecTypes['pg/int8@1']['output'] | null;
      readonly bInt2: CodecTypes['pg/int8@1']['output'] | null;
      readonly float: CodecTypes['pg/float8@1']['output'] | null;
      readonly float2: CodecTypes['pg/float8@1']['output'] | null;
      readonly bytes: CodecTypes['pg/bytea@1']['output'] | null;
      readonly bytes2: CodecTypes['pg/bytea@1']['output'] | null;
      readonly bool: CodecTypes['pg/bool@1']['output'] | null;
      readonly bool2: CodecTypes['pg/bool@1']['output'] | null;
      readonly dt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly dt2: CodecTypes['pg/timestamptz@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly TestModel: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly string: CodecTypes['pg/text@1']['input'] | null;
      readonly string2: CodecTypes['pg/text@1']['input'] | null;
      readonly int: CodecTypes['pg/int4@1']['input'] | null;
      readonly int2: CodecTypes['pg/int4@1']['input'] | null;
      readonly bInt: CodecTypes['pg/int8@1']['input'] | null;
      readonly bInt2: CodecTypes['pg/int8@1']['input'] | null;
      readonly float: CodecTypes['pg/float8@1']['input'] | null;
      readonly float2: CodecTypes['pg/float8@1']['input'] | null;
      readonly bytes: CodecTypes['pg/bytea@1']['input'] | null;
      readonly bytes2: CodecTypes['pg/bytea@1']['input'] | null;
      readonly bool: CodecTypes['pg/bool@1']['input'] | null;
      readonly bool2: CodecTypes['pg/bool@1']['input'] | null;
      readonly dt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly dt2: CodecTypes['pg/timestamptz@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly testModel: {
      readonly bInt: CodecTypes['pg/int8@1']['output'] | null;
      readonly bInt2: CodecTypes['pg/int8@1']['output'] | null;
      readonly bool: CodecTypes['pg/bool@1']['output'] | null;
      readonly bool2: CodecTypes['pg/bool@1']['output'] | null;
      readonly bytes: CodecTypes['pg/bytea@1']['output'] | null;
      readonly bytes2: CodecTypes['pg/bytea@1']['output'] | null;
      readonly dt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly dt2: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly float: CodecTypes['pg/float8@1']['output'] | null;
      readonly float2: CodecTypes['pg/float8@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly int: CodecTypes['pg/int4@1']['output'] | null;
      readonly int2: CodecTypes['pg/int4@1']['output'] | null;
      readonly string: CodecTypes['pg/text@1']['output'] | null;
      readonly string2: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly testModel: {
      readonly bInt: CodecTypes['pg/int8@1']['input'] | null;
      readonly bInt2: CodecTypes['pg/int8@1']['input'] | null;
      readonly bool: CodecTypes['pg/bool@1']['input'] | null;
      readonly bool2: CodecTypes['pg/bool@1']['input'] | null;
      readonly bytes: CodecTypes['pg/bytea@1']['input'] | null;
      readonly bytes2: CodecTypes['pg/bytea@1']['input'] | null;
      readonly dt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly dt2: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly float: CodecTypes['pg/float8@1']['input'] | null;
      readonly float2: CodecTypes['pg/float8@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly int: CodecTypes['pg/int4@1']['input'] | null;
      readonly int2: CodecTypes['pg/int4@1']['input'] | null;
      readonly string: CodecTypes['pg/text@1']['input'] | null;
      readonly string2: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;